refactor: Extract markdown parsing
We need to create data and metadata objects from the markdown nodes.
To keep this organized, it's easier to separate the data-parsing code
from the code that produced the structure of hast from mdast.
Basically: parse the data separately form the structure of the presentation
